New implementation of osSignal
Removed thread termination signal, because the user can implement it himself if used.

Implemented signal so that it can fit more perfectly for waking up a driver thread from an interrupt. The interrupt can now post a information block regarding the peripheral status details to the thread waiting for a specific signal.